+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [1.1.0] - 2018-06-27
+
+### Added
+
++ Added language server for code completion and on-demand documentation on
+  hover
++ Added snippets
+
 ## [1.0.3] - 2018-06-23
 
 ### Removed
@@ -40,7 +48,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 + Initial release
 
-[Unreleased]: https://github.com/mserajnik/dnh/compare/1.0.3...develop
+[Unreleased]: https://github.com/mserajnik/dnh/compare/1.1.0...develop
+[1.1.0]: https://github.com/mserajnik/dnh/compare/1.0.3...1.1.0
 [1.0.3]: https://github.com/mserajnik/dnh/compare/1.0.2...1.0.3
 [1.0.2]: https://github.com/mserajnik/dnh/compare/1.0.1...1.0.2
 [1.0.1]: https://github.com/mserajnik/dnh/compare/1.0.0...1.0.1